change/fix library header location
Previously I had been putting the library header files in a generic
"local/include" location then using the CPATH environment variable to point to
this location when building my small projects. But at some point I can imagine
wanting to create a set of cross-development tools, each of which will target
a separate platform. Anticipating this change I put, instead, the header files
associated with this specific platform in a location that is relevant to the
build tools for this specific board.

Therefore in the future if I do end up with multiple cross-development tools,
the header files for each set of tools will be specific to each set of tools.

This is already what I'm doing for the library files, so I'm just following
suit with the header files too.

This change eliminates the need to setup the CPATH environment variable.